Extracts field builder so that mutations can share the logic
